bluez: Add bluez-daemon package
Move bluetoothd and related programs and files from bluez-utils into
bluez-daemon package. The main motivation for this is to remove the dependency
to DBus from bluez-utils. The utilities like "hcitool" are useful without it.

Also add kmod-bluetooth dependency to bluez-utils, as the utils don't make much
sense without bluetooth support in the kernel.

Change-Id: I852991b0b5aeb734e0fdf2788e48c5d02a01a3b6
Signed-off-by: Bruno Randolf <br1@einfach.org>
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>