python-six: new package
From the README:

Six is a Python 2 and 3 compatibility library.  It provides utility functions
for smoothing over the differences between the Python versions with the goal of
writing Python code that is compatible on both Python versions.  See the
documentation for more information on what is provided.

Signed-off-by: Jeffery To <jeffery.to@gmail.com>
(cherry picked from commit a9af921700b5404e56beb723b6a77bfbc4018200)
Signed-off-by: Sandhya <sandhyks@codeaurora.org>

Change-Id: I23251511ec82900cc071d843979bf1293ba30a89
Signed-off-by: Sandhya <sandhyks@codeaurora.org>