python: Fix host python build with ubuntu 32bit system
Currently we are getting compilation errors when python is built
on ubuntu 32 bit host build system. Python build tries to take
ffi library which is part of python source but it is failing to
compile with ubuntu 32bit system.

In order to fix this, we will use openwrt ffi library which will
ensure that python is linked with proper ffi library.

This will also help to ensure that our tools are linked with the
same version of libffi library.

Change-Id: I283e59a9a832688c1f03885ddf409b86fb84386d
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>