multiwan: fixes local route table is not in the rule list
ip rule flush command will delete all of entries in the rule list
except of 0 priority.
Usually, the local route table should be 0 priority in the ip rules,
but this is not the case in the new version openwrt

check if the local route talbe exists in the rule list,
add the local route table to ip rule list when it doesn't exist.

Change-Id: Ia1bcad4b3cc177dbbc797e5180f9eaaa20ebb1d3
Signed-off-by: Zhu Ken <guigenz@codeaurora.org>