openswitch: improve build time for kmod-only builds
When only kmod-openvswitch is enabled, which commonly happens during LEDE
release builds, there is no need to build the entire userland of openvswitch
as this is done at a later stage on systems dedicated to build only userspace
packages.

This change conditionalizes the dependency on python, which means that python
prerequisites are only compiled if the openvswitch-python package is enabled,
which allows us to eliminate the entire python dependency chain.

Furthermore, this change sets MAKE_PATH to the kernel module sub directory
if only kmod-openvswitch is enabled which causes the openvswitch build system
to only process sources related to the .ko files.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>
(cherry picked from commit bf7ef2af120914598aa9b177ab0960a605c214f0)
Signed-off-by: Sandhya <sandhyks@codeaurora.org>

Change-Id: I29bdd384a00544ec8528d977df38c8abe4e04d1e
Signed-off-by: Sandhya <sandhyks@codeaurora.org>