unbound: expand UCI support for odhcpd DHCP-DNS
This is bare minimum change in 'unbound.sh' and
'dnsmasq.sh' to migrate the UCI option set for
more flexibility. The boolean(s) to link to
dnsmasq are being changed to a state to include
odhcpd. It is executable but a small step for
clear change management.

Signed-off-by: Eric Luehrsen <ericluehrsen@hotmail.com>
(cherry picked from commit 89e7bb8f38cbe9bec71b2d88a63e346230ed29c7)

Change-Id: Id979267cfcc6b0d977616e5f119ed32b1d818dc9
Signed-off-by: Sanjeev N <nsanjeev@codeaurora.org>
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>