openvswitch: enforce the use of the python binary from `packages`
Seems it was not failing for me because it was probably
using my host Python, which may have the `six` package
available.

This patch enforces the use of the packaged Python.
That way, it's more consistent that the python-six
package is available.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>
(cherry picked from commit e26a34c6b4f634240d2dd562efb8649654006b23)
Signed-off-by: Sandhya <sandhyks@codeaurora.org>

Change-Id: I848cfaa03a99a69538d0eab601639745d66d86f8
Signed-off-by: Sandhya <sandhyks@codeaurora.org>