xl2tpd: Do not use proto_add_dependency for L2TP over IPSec
When L2TP is established over IPSec, we add a route to the LNS
using the ipsec interface before L2TP connection is established.
However, l2tp.sh is overriding this route with a route through
the physical WAN interface when L2TP tries to come up. This
fix is to avoid overriding of the IPSec route when l2TP is coming
up in the case of L2TP over IPSec.

Change-Id: Ic6932ad4e70838717e0e05049eafcc732bbc2a34
Signed-off-by: Gaurav Pathak <gaurpath@codeaurora.org>