MQTT for AWS: Customisation of MQTT package for IoT Control
This change enables additonal support from the mosquitto package to
to enable Iot control of devices using Amazon's AWS Cloud via MQTT
protocol.This is done by adding an option '-iot' to the application.
This provides support to use case/scenarios when the voice commands
are sent from one gateway, and the Iot devices are connected to
another gateway in a different location.
For Ex- To control lights at home by sending voice commands from office.
This could be done with the Voice Control Gateway at the office, and
commands are sent to the cloud and the shadow update in the cloud is
sent to the gateway at home using MQTT to control the IoT devices
at home using iotutils.

Currently smart home devices are controlled over HTTPS using curl,
or by MQTT with Websockets as in AVS application. Moving to MQTT
makes the communication lightweight.

When 'mosquitto_sub' command is launched, the device/mosquitto
application subscribes to the AWS topic as mentioned in the
command  and listens to the cloud for any delta in the shadow
state of the AWS IoT thing. Any shadow update is communicated to
the device through port 8883. When the '-iot' option is not used
with the 'mosquitto_sub' command,the application just subscribes
to the topic and keeps listening for updates without any further action.

When -iot is used in the command,it parses the JSON data received
from the subscription, extracts the parameters and invokes the
'iotutils' application with the necessary arguments. The new state
value is published/reported  back to the device shadow in the cloud
on the parent topic.

Change-Id: I2d8055b678dc934a79c239ad87c2c21e2ad10875
Signed-off-by: Sandhya K S <sandhyks@codeaurora.org>