xtables-addons: pass correct flags to compile and install
The Makefile currently redefine the Compile and Install functions.
This is not working when using an external toolchain because some
flags are not interpreted, like CROSS_COMPILE. It is possible to
override the MAKE_FLAGS and MAKE_INSTALL_FLAGS instead.

Signed-off-by: Sébastien Blin <sebastien.blin@savoirfairelinux.com>
(cherry picked from commit 32aaaaa7d37980de7482a49a729009cad6b3e28d)

Change-Id: Ie879cd372680ebd10c431d8853957ec1d7cde8b8
Signed-off-by: Ronghua Zeng <quic_ronghuaz@quicinc.com>